Added Trees to create a static scene

+++ GraphX-Rendering-Engine/src/entities/Terrain.cpp
@@ -53,7 +53,7 @@ namespace engine
 			{
 				// Calculate the vertices of the terrain
 				double yCoord = GetYCoords(x, z);
-				vertex.Position = gm::Vector3(x * m_TileSize, (float)yCoord, -z * m_TileSize);
+				vertex.Position = gm::Vector3(x * m_TileSize, yCoord, -z * m_TileSize);
 				vertex.Normal = CalculateNormal(x, z);	// TODO: Optimize using another loop
 				vertex.TexCoord = gm::Vector2((float)x, (float)z);
 				m_Vertices->emplace_back(vertex);
@@ -120,13 +120,13 @@ namespace engine
 		return Interpolate(i1, i2, fracZ);
 	}
 
-	gm::Vector3 Terrain::CalculateNormal(double x, double z)
+	gm::Vector3 Terrain::CalculateNormal(int x, int z)
 	{
 		double heightL = GetYCoords(x - 1, z);
 		double heightR = GetYCoords(x + 1, z);
 		double heightD = GetYCoords(x, z - 1);
 		double heightU = GetYCoords(x, z + 1);
-		gm::Vector3 normal = gm::Vector3(heightL - heightR, 2.0, heightD - heightU);
+		gm::Vector3 normal = gm::Vector3((float)(heightL - heightR), 2.0, (float)(heightD - heightU));
 		normal.Normalize();
 		return normal;
 	}
